helper-apps/cortex-autogen2/requirements.txt
# -------------------------------------------------------------------
# ESSENTIAL – office docs, PDFs, images
# -------------------------------------------------------------------
Pillow>=11.0.0
python-magic>=0.4.27
chardet>=5.2.0
PyPDF2>=3.0.1
pdfminer.six>=20221105
docx2txt>=0.9,<1.0
openpyxl>=3.1.2
xlrd>=2.0.1
python-pptx>=0.6.23
odfpy>=1.4.1
pdf2image>=1.16.3
python-pptx-interface>=0.0.14

# -------------------------------------------------------------------
# MEDIA – images / audio / video
# -------------------------------------------------------------------
opencv-python>=4.9.0
imageio[ffmpeg]>=2.34.0
CairoSVG>=2.7.1
pyheif>=0.6.3
moviepy>=1.0.3
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1

# -------------------------------------------------------------------
# ARCHIVES
# -------------------------------------------------------------------
patool>=1.12
py7zr>=0.21.0
rarfile>=4.0

# -------------------------------------------------------------------
# DATA & SCIENTIFIC
# -------------------------------------------------------------------
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.14.0
polars>=1.8.2
pyarrow>=16.1.0
tabula-py>=2.8.1
h5py>=3.10.0
netCDF4>=1.6.5
ezdxf>=1.1.1
seaborn>=0.13.0
plotly>=5.24.1
kaleido>=0.2.1
wordcloud>=1.9.3
yfinance>=0.2.0
tabulate>=0.9.0

# -------------------------------------------------------------------
# NLP & TEXT PROCESSING
# -------------------------------------------------------------------
nltk>=3.9.1
spacy>=3.8.4

# -------------------------------------------------------------------
# ARABIC & RTL TEXT SUPPORT
# -------------------------------------------------------------------
arabic-reshaper>=3.0.0
python-bidi>=0.4.2

# -------------------------------------------------------------------
# EMAIL / WEB / E-BOOK
# -------------------------------------------------------------------
eml-parser>=1.15.0
msg-parser>=1.2.0,<2.0
beautifulsoup4>=4.12.3
html5lib>=1.1
ebooklib>=0.18

# -------------------------------------------------------------------
# CLOUD / INFRASTRUCTURE
# -------------------------------------------------------------------
azure-functions>=1.20.0
azure-storage-blob>=12.19.0
azure-storage-queue>=12.10.1
redis>=5.0.4
sqlalchemy>=2.0.30
pymysql>=1.1.1
requests>=2.32.3
aiohttp>=3.12.14
aiofiles>=24.1.0

# -------------------------------------------------------------------
# AI / AUTOMATION
# -------------------------------------------------------------------
openai>=1.97.1
tiktoken>=0.9.0
autogen-agentchat[openai]>=0.7.5
autogen-ext[docker]>=0.7.5
docker>=7.1.0
playwright>=1.54.0
markitdown>=0.1.2
python-dotenv>=1.0.1
reportlab>=4.2.5
fpdf2>=2.7.9
matplotlib>=3.9.0

# -------------------------------------------------------------------
# OPTIONAL EXTRAS
# -------------------------------------------------------------------
imageio-ffmpeg>=0.4.9
python-lzf>=0.2.4
